create-kpatch-module: Fix -Wconversion warnings
Fix warnings reported by -Wconversion,-Wno-sign-conversion CFLAGS.

+++ kmod/patch/kpatch-patch.h
@@ -40,7 +40,7 @@ struct kpatch_patch_dynrela {
 	char *name;
 	char *objname;
 	int external;
-	int addend;
+	long addend;
 };
 
 struct kpatch_pre_patch_callback {
